Create API public of the render
